backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0
cryptography==41.0.7 